% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_census_data.R
\name{city_census_data}
\alias{city_census_data}
\title{Intersect census data with boundaries}
\usage{
city_census_data(census_var, bounds)
}
\arguments{
\item{census_var}{table with the census variable of interest and associated geometries
can use clean_census_data to get this input

(created for census tracts & city boundaries, should work for other sizes)}

\item{bounds}{the boundaries to intersect with the census variable}
}
\value{
city_int -- the intersected geometries with the boundary
}
\description{
Intersect census data with boundaries
}
\examples{




}
