% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_census_data.R
\name{clean_census_data}
\alias{clean_census_data}
\title{Cleaning Census data -- removing all but relevant variable and geometries}
\usage{
clean_census_data(census_table, variable_n)
}
\arguments{
\item{census_table}{the table of census values extracted with API key

 example -- pulling in census data: it should look something like this:

 mpls_income2019 <- get_acs(geography = "tract", year = 2019,
 variables = c(med_earn = "S2001_C01_002E"), state = "MN",
 county = "Hennepin", survey = "acs5", output = "wide",
 geometry = TRUE)}

\item{variable_n}{the variable of interest, should be a character string}
}
\value{
a clean table with only the census variable and its geometries
Note: can be used int eh city_census_data function
}
\description{
Cleaning Census data -- removing all but relevant variable and geometries
}
